Most of the Skeleton code docs (#3)
* DC motor draft init (not yet the example in .ino)

Include all the API in the motorcontrol.hpp and cpp.
(need to add the example of how to use the API in the .ino file (main.cpp))

* Adding the documentation of sensor

Added sensors: IMU (ICM42668-P), Ultrasonic sensor (HC-04), RFID Reader (MRFC522) and Buzzer
Not yet do: the how to call the APL, such as create the obj of the IMU and RFID reader in .ino file.

* .ino md init

added movement and .ino md
Not yet do: examaple of RFID and IMU use in .ino file

* fix: heading display

* fix: enlarge image width

* fix: fix img path and remove unused img

* fix: styling and link

* fix: format links

* feat: firebase part done

---------

Co-authored-by: Jason Chan <jason.mc.chan@gmail.com>